+++ lib/+mp/mme_buslink_opf.m
@@ -7,8 +7,9 @@ classdef (Abstract) mme_buslink_opf < mp.mme_buslink
 % Implements (currently empty) method for forming an interior initial point.
 
 %   MATPOWER
-%   Copyright (c) 2022-2024, Power Systems Engineering Research Center (PSERC)
+%   Copyright (c) 2022-2026, Power Systems Engineering Research Center (PSERC)
 %   by Ray Zimmerman, PSERC Cornell
+%   and Wilson Gonzalez Vanegas, Universidad Nacional de Colombia Sede Manizales
 %
 %   This file is part of MATPOWER.
 %   Covered by the 3-clause BSD License (see LICENSE file for details).
@@ -21,8 +22,10 @@ classdef (Abstract) mme_buslink_opf < mp.mme_buslink
         function x0 = interior_x0(obj, mm, nm, dm, x0)
             %
         end
-        
+
         function obj = data_model_update_on(obj, mm, nm, dm, mpopt)
+            %
+
             dme = obj.data_model_element(dm);
             nme = obj.network_model_element(nm);
 
@@ -30,11 +33,11 @@ classdef (Abstract) mme_buslink_opf < mp.mme_buslink
 
             for p = 1:nme.nz
                 %% buslink complex power flows
-                zbl = nm.soln.z(ss.i1.buslink(p):ss.iN.buslink(p));                
+                zbl = nm.soln.z(ss.i1.buslink(p):ss.iN.buslink(p));
 
                 %% update in the data model
-                dme.tab.(sprintf('pg%d_start', p))(dme.on) = real(zbl) * dm.base_kva;
-                dme.tab.(sprintf('qg%d_start', p))(dme.on) = imag(zbl) * dm.base_kva;
+                dme.tab.(sprintf('p%d', p))(dme.on) = real(zbl) * dm.base_kva;
+                dme.tab.(sprintf('q%d', p))(dme.on) = imag(zbl) * dm.base_kva;
             end
         end
     end     %% methods